+++ R/package-init.R
@@ -1,5 +1,7 @@
 #' @import jsonlite
-#' @import httr
+#' @import httr2
+#' @importFrom httr GET content add_headers
+#' @importFrom curl has_internet
 #' @importFrom graphics par plot.new rasterImage
 #' @importFrom stringr str_remove str_detect
 #' @importFrom grDevices as.raster