+++ NEWS.md
@@ -1,3 +1,6 @@
-# peermodels 0.10.2
+# peermodels 0.10.3
 
-* new release.
+## New Features:
+* Results from asynchronous requests can now be emailed back to the user.
+* Switched to httr2 for better error management
+* Requests fail gracefully when the server is down.